+++ graphql/resolvers/group.js
@@ -8,18 +8,30 @@ const groupResolver = {
   },
 
   // ----- MUTATIONS -----
-  createGroup: async (args) => {
+  createGroup: async (args,req) => {
+    if (!req.isAuth) {
+      throw new Error("unauthenticated");
+      
+    }
     const group = new Group({ name: args.groupInput.name });
     await group.save();
     return group;
   },
 
-  UpdateGroup: async (args) => {
+  UpdateGroup: async (args,req) => {
+    if (!req.isAuth) {
+      throw new Error("unauthenticated");
+      
+    }
     const { id, groupInput } = args;
     return await Group.findByIdAndUpdate(id, { name: groupInput.name }, { new: true });
   },
 
-  DeleteGroup: async (args) => {
+  DeleteGroup: async (args,req) => {
+    if (!req.isAuth) {
+      throw new Error("unauthenticated");
+      
+    }
     const { id } = args;
     await Group.findByIdAndDelete(id);
     return true;
